Extract get_stories to stories helper
